Belated commit
A number of new scripts added that I used to create figures for
my AGU poster.
Also added support scripts to morph weather files from morphing
factors that are pre-computed instead of done when the script
is run. There is also a means of produced formatted excel tables
with statistics from the morphed weather files.